+++ ingenialink/dictionary.py
@@ -212,8 +212,6 @@ class Dictionary(ABC):
     """Instance of all the errors in the dictionary."""
     image: Optional[str] = None
     """Drive's encoded image."""
-    moco_image: Optional[str] = None  # TODO study COM-KIT case
-    """Motion CORE encoded image. Only available when using a COM-KIT."""
     is_safe: bool = False
     """True if has SafetyPDOs element, else False"""
     _registers: Dict[int, Dict[str, Register]]
@@ -238,6 +236,26 @@ class Dictionary(ABC):
             self.read_dictionary()
         except KeyError as e:
             raise ILDictionaryParseError("The dictionary is not well-formed.") from e
+
+    def __add__(self, other_dict: "Dictionary") -> "Dictionary":
+        """Merge two dictionary instances.
+
+        It can only be used for merging COM-KIT and CORE dictionaries.
+
+        """
+        if not isinstance(other_dict, type(self)):
+            raise TypeError(
+                f"Cannot merge dictionaries. Expected type: {type(self)}, got: {type(other_dict)}"
+            )
+        if not other_dict.is_coco_dictionary and not self.is_coco_dictionary:
+            raise ValueError(
+                "Cannot merge dictionaries. One of the dictionaries must be a COM-KIT dictionary."
+            )
+        self._merge_registers(other_dict)
+        self._merge_errors(other_dict)
+        self._merge_attributes(other_dict)
+        self._set_image(other_dict)
+        return self
 
     def registers(self, subnode: int) -> Dict[str, Register]:
         """Gets the register dictionary to the targeted subnode.
@@ -313,6 +331,61 @@ class Dictionary(ABC):
         if uid in self.safety_tpdos:
             return self.safety_tpdos[uid]
         raise KeyError(f"Safe TPDO {uid} not exist")
+
+    def _merge_registers(self, other_dict: "Dictionary") -> None:
+        """Add the registers from another dictionary to the dictionary instance.
+
+        Args:
+            other_dict: The other dictionary instance.
+
+        """
+        for subnode, registers in other_dict._registers.items():
+            self._registers[subnode].update(registers)
+
+    def _merge_errors(self, other_dict: "Dictionary") -> None:
+        """Add the errors from another dictionary to the dictionary instance.
+
+        Args:
+            other_dict: The other dictionary instance.
+
+        """
+        self.errors.errors.update(other_dict.errors.errors)
+
+    def _set_image(self, other_dict: "Dictionary") -> None:
+        """Set the image attribute.
+
+        Choose the image from the dictionary that has one.
+
+        Args:
+            other_dict: The other dictionary instance.
+
+        """
+        core_dict = self if other_dict.is_coco_dictionary else other_dict
+        self.image = core_dict.image
+
+    def _merge_attributes(self, other_dict: "Dictionary") -> None:
+        """Add the revision number, product code, firmware version and part number
+        from the other dictionary to the dictionary instance.
+
+        Args:
+            other_dict: The other dictionary instance.
+
+        """
+        if not other_dict.is_coco_dictionary:
+            self.product_code = other_dict.product_code
+            self.revision_number = other_dict.revision_number
+            self.firmware_version = other_dict.firmware_version
+            self.part_number = other_dict.part_number
+
+    @property
+    def is_coco_dictionary(self) -> bool:
+        """Check if dictionary is a CoCo dictionary
+
+        Returns:
+            True if the dictionary is a CoCo dictionary. False otherwise.
+
+        """
+        return len(self.registers(1)) == 0
 
 
 class DictionaryV3(Dictionary):
@@ -895,7 +968,6 @@ class DictionaryV2(Dictionary):
     DICT_ENUMERATIONS = "./Enumerations"
     DICT_ENUMERATIONS_ENUMERATION = f"{DICT_ENUMERATIONS}/Enum"
     DICT_IMAGE = "DriveImage"
-    DICT_MOCO_IMAGE_ATTRIB = "moco"
 
     dict_interface: Optional[str]
 
@@ -967,16 +1039,9 @@ class DictionaryV2(Dictionary):
                 if current_read_register:
                     self._add_register_list(current_read_register)
         try:
-            images = root.findall(self.DICT_IMAGE)
-            for image in images:
-                if image.text is not None and image.text.strip():
-                    if (
-                        "type" in image.attrib
-                        and image.attrib["type"] == self.DICT_MOCO_IMAGE_ATTRIB
-                    ):
-                        self.moco_image = image.text
-                    else:
-                        self.image = image.text
+            image = root.find(self.DICT_IMAGE)
+            if image is not None and image.text is not None and image.text.strip():
+                self.image = image.text
         except AttributeError:
             logger.error(f"Dictionary {Path(self.path).name} has no image section.")
         # Closing xdf file