+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Dictionary V3 support
 - ILWrongWorkingCount exception raised when the working count in PDO is not the expected one.
 - Register type for monitoring/disturbance data.
+- Support for merging dictionary instances. (It can only be used for merging COM-KIT and CORE dictionaries.)
 
 ### Changed
 - Dictionary class properties, such as subnodes and interface.